Added some fixes

+++ js/__app/createCall.js
@@ -48,8 +48,14 @@
                        },
                        error: function (error, xhr)
                        {
-                           $(location).attr('href', 'login.php');
-                           localStorage.clear();
+                           if (error == 'Unauthorized') {
+                               $(location).attr('href', 'login.php');
+                               localStorage.clear();
+                               alert('Session expired');
+                           } else {
+                               alert(xhr.responseText);
+                           }
+
                        },
                        dataType: 'json',
                        success: function (data)
